builder/agent_skills/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64.0.0", "wheel"]

[project]
name = "agent_skills"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[redis,mcp,langchain,phoenix] @ git+https://github.com/NVIDIA/NeMo-Agent-Toolkit.git@v1.5.0",
  "pyyaml>=6.0",
]
requires-python = ">=3.11,<3.14"
description = "Agent Skills loader for NeMo Agent Toolkit — discovers, loads, and executes Anthropic-style SKILL.md skill directories."
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
agent_skills = "agent_skills.register"

[tool.setuptools]
packages = ["agent_skills"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
agent_skills = ["configs/*.yml", "configs/*.yaml"]

[tool.uv]
config-settings = { editable_mode = "compat" }
